Replace GMP with MbedTLS
Libbyzea's State class currently uses GMP for multi-precision integer
calculations. This is not available on all plattforms and MbedTLS also
comes with a multi-precision integer library. Since we already depend on
MbedTLS it makes sense to replace usages of GMP with the multi-precision
integer library from MbedTLS. This commit does exactly that.