Cycle_counter: add ESP32-C3 support
The Cycle_counter class is used in various places to measure time in the
form of processor clock cycles. This functionality is implemented using
Intel's rdtsc instruction. This instruction is not available on the
RISC-V-based ESP32-C3. In order to support this processor, it is
necessary to use performance counters available there.

This commit adds a hardware-abstraction namespace called 'platform' that
provides functions to get the value of a cycle counter and its maximum
value. Additionally, it adjusts the Cycle_counter to use this
hardware-abstraction in order to make it compatible with the ESP32-C3.