Fix bad new_key debug message
The code contains a debug message printed when receiving bad new_key
messages from replicas. This message is also printed for the replica's
own new_key messages, which are never verified as a special case of the
verify function. Therefore, the error message is misleading. This commit
removes the debug output when the new_key message was sent from the
replica to itself.