State: fix state transfer not working
Commit e6c5fb9b swaped GMP's multi-precision integers with the MbedTLS
implementation in order to reduce the number of dependencies. The
commit's changes introduced a bug in PBFT's original digest generation
algorithm, since it used only a part of an input buffer to the digest
algorithm. On the code checking digests for the state transfer, the
whole buffer was read, leading to different digests for the same digest
sum.

This commit fixes the issue, by letting the regular digest generation
read the full buffer, instead of the partial content, so that digest
generation behaves the same in every part of the State class.