Status: fix assertion triggering for valid replicas
The Status message holds a bitmap with a bit for each replica from which
the sending replica has received a New_view message for in the current
view. The code handling setting and testing this bitmap asserts that the
given replica is in rage of the bitmap. The code compares the replica id
with the size of the bitmap in bytes, instead of the size in bits. This
commit fixes the issue so the assertion is not triggered for valid
replicas.