View_change: remove window size assert
The View_change message has an assert that enforces a minimum window
size of 64. Actually, it creates a bitmap for holding information
whether a certain request has been prepared or not on the sender
replica. This bitmap is based on an integer array and the assert forces
its size to by divisible by two. In my opinion, there does not seem to
be a reason why this has to be strictly necessary. Perhaps it is for
alignment reasons. This commit removes the assert, which allows us to
use window sizes smaller than 64, which can drastically reduce memory
consumption of the library.